Introduce SQLModel
SQLModel is a wrapper on top of SQLAlchemy and Pydantic.
https://sqlmodel.tiangolo.com/features/

Model schemas and metadatas at both the database level and the API
response level.

The read API responses should be unchanged by this, since SQLModel is
a layer on top of Pydantic.

The database models differ from these in a couple of ways:
- column definitions and tags are defined as JSON fields
- there are integer primary key columns