Introduce SQLModel
SQLModel is a wrapper on top of SQLAlchemy and Pydantic.
https://sqlmodel.tiangolo.com/features/

It allows us to create model classes using the same syntax that can work
for both request/response serialization/deserialization and
object-relational mapping (ORM). Using an ORM framework like sqlalchemy
means we can easily persist metadata information to a database that uses
SQL, without having to work the SQL directly.

We haven't yet replaced the S3 data store, but I've introduced some
tests that use an in-memory sqlite database to demonstrate how this
might work.

The different repsonse models will include different combinations of
fields, e.g. when reading a data product, we can include the ID,
version, and other generared fields.

The database models also need to differ from the response models:
- column definitions and tags are defined as JSON fields in SQLAlchemy,
  but elsewhere we use a nested Column model (this is clunky)
- internally we have integer primary key columns, but externally we have
  defined a human friendly ID (dp:my-data-product:version)
- we have defined several other metadata fields we will need to store,
  but we haven't implemented these yet

Some gotchas and annoyances:
- The Field.regex attribute doesn't get propagated to the JSON schema(!?)
- I had trouble defining Columns as a nested model in the table model,
  while treating it as a SQLAlchemy JSON field
- There is currently no reference documentation
- Some knowledge of SQLAlchemy and Pydantic is assumed to use SQLModel

If we decide not to proceed with SQLModel, then the *Table models can be replaced
with SQLAlchemy, and the rest of the models can be replaced with
Pydantic again.

If we decide not to use SQL for our backing database, then we should
also revert this change.