Remove version number from IDs in requests
Addressing DataProductVersions by ID in the API didn't work, because when
you update a data product/schema, or delete a schema, you actually
create a new version of the data product.

In such situations, it would be really weird for the resource in the
response to have a different ID to the one the user requested.

From now on, we will refer to different versions of the data product
with the same ID, and have the API always return the latest version.